+++ src/index.ts
@@ -1,1 +1,3 @@
 export {transform} from './transform';
+export {Reporter} from './reporter';
+export {create_runtime as setup} from './runtime';