+++ src/transform.ts
@@ -1,31 +1,30 @@
 import * as path from 'path';
 import * as ts from 'typescript';
 import {
+  config_namespace,
+  package_name,
   runtime_namespace,
   trigger_regex,
+  Expressions,
   TestMethod,
   Trigger,
-  TriggerDescriptions,
   TriggerFlag,
+  TriggerLines,
   TriggerMatchArray,
   TriggerMatchIndex,
 } from './definitions';
-
-// tslint:disable-next-line:no-require-imports no-var-requires
-const package_name = require('../package.json').name;
+import {defaults, remove_spaces, traverse_node} from './utils';
 
 export const transform = (source_text: string, source_filename: string, _jest_config?: any) => {
   const source_file = ts.createSourceFile(source_filename, source_text, ts.ScriptTarget.Latest, false);
 
   const triggers = create_triggers(source_file);
-  const trigger_lines = triggers.map(trigger => trigger.line);
-
   return triggers.reduce<string[]>(
     (transformed_line_texts, trigger) => {
       transformed_line_texts[trigger.line] += `${create_test_expression(trigger)};`;
       return transformed_line_texts;
     },
-    source_text.split('\n').map((_, index) => (index !== 0) ? '' : `${create_setup_expression(trigger_lines)};`),
+    source_text.split('\n').map((_, index) => (index !== 0) ? '' : `${create_setup_expression(triggers)};`),
   ).join('\n');
 };
 
@@ -82,7 +81,7 @@ function create_triggers(source_file: ts.SourceFile): Trigger[] {
 
     if (trigger_line in partial_triggers) {
       try {
-        const expression = node.getText(source_file);
+        const expression = node.getText(source_file).replace(/\s*;?\s*$/, '');
 
         const partial_trigger = partial_triggers[trigger_line];
         delete partial_triggers[trigger_line];
@@ -103,7 +102,7 @@ function create_triggers(source_file: ts.SourceFile): Trigger[] {
     const relative_filename = path.relative(process.cwd(), source_file.fileName);
     throw new Error(`Unattachable trigger(s) detected:\n\n${
       unattachable_lines
-        .map(line => `  ${relative_filename}:${line + 1} ${partial_triggers[line].description || ''}`)
+        .map(line => `  ${relative_filename}:${line + 1} ${defaults(partial_triggers[line].description, '')}`)
         .join('\n')
         .replace(/\s+$/mg, '')
     }`);
@@ -112,20 +111,30 @@ function create_triggers(source_file: ts.SourceFile): Trigger[] {
   return triggers;
 }
 
-function traverse_node(node: ts.Node, callback: (node: ts.Node) => void) {
-  if (node.kind !== ts.SyntaxKind.SourceFile) {
-    callback(node);
-  }
-  ts.forEachChild(node, child => traverse_node(child, callback));
-}
+function create_setup_expression(triggers: Trigger[]) {
+  const expressions = triggers.reduce<Expressions>(
+    (infos, trigger) => ({
+      ...infos,
+      [trigger.line]: trigger.expression,
+    }),
+    {},
+  );
 
-function create_setup_expression(trigger_lines: number[]) {
   const stringify_package_name = JSON.stringify(package_name);
-  const stringify_trigger_lines = JSON.stringify(trigger_lines);
+  const stringify_expressions = JSON.stringify(expressions);
+  const config_expression = remove_spaces(`
+    (function () {
+      try {
+        return ${config_namespace};
+      } catch (e) {
+        return {};
+      }
+    })()
+  `);
   return remove_spaces(`
     var ${runtime_namespace};
     beforeAll(function (done) {
-      require(${stringify_package_name}).setup(module.filename, ${stringify_trigger_lines}, function () {
+      require(${stringify_package_name}).setup(module, ${config_expression}, ${stringify_expressions}, function () {
         ${runtime_namespace} = arguments[0];
         done();
       });
@@ -134,18 +143,16 @@ function create_setup_expression(trigger_lines: number[]) {
 }
 
 function create_test_expression(trigger: Trigger) {
-  const stringify_description = JSON.stringify(trigger.description || trigger.expression);
+  const description = defaults(trigger.description, trigger.expression).replace(/\n/g, '\n    ');
+  const stringify_description = JSON.stringify(description);
+  const report_expression = `${runtime_namespace}.report(${trigger.line})`;
   const snapshot_expression = `${runtime_namespace}.snapshot(${trigger.line})`;
   const assertion_expression = (trigger.flag === TriggerFlag.Show)
-    ? `console.log(${snapshot_expression})`
+    ? `console.log(${report_expression})`
     : `expect(${snapshot_expression}).toMatchSnapshot()`;
   return remove_spaces(`
     ${trigger.method}(${stringify_description}, function () {
       ${assertion_expression};
     })
   `);
-}
-
-function remove_spaces(text: string) {
-  return text.replace(/^\s+|\s+$|\n/mg, '');
 }